Refactor/should bot message be muzzled (#125)
* Refactor phase 1 of shouldBotMessageBeMuzzled

* Added userIdByBlocksSpoiler to designate spoiler specific userIds

* Added logs to getBotId

* Changed getUserId to return undefined when a user is not found

* Fixed returning an empty string from getUserId and removed unnecessary logs

* Added log for muzzles inside of shouldBackfire due to suspicions around max percent change

* Added caching to getAllUsers and added logs of each request

* Fixed import

* Added better backfireId handling

* Added fixes for type coercion regarding string to numbers

* One more type coercion fix for sendMuzzledMessge

* cleaned up ugly code in reaction.persistence

* Added catchs to webClient interactions to avoid JP from breaking things

* Added ? to request.event.blocks.length in shouldBotMessageBeMuzzled

+++ src/controllers/event.controller.ts
@@ -12,7 +12,6 @@ import { ReactionService } from '../services/reaction/reaction.service';
 import { SentimentService } from '../services/sentiment/sentiment.service';
 import { SlackService } from '../services/slack/slack.service';
 import { WebService } from '../services/web/web.service';
-import { CounterMuzzle } from '../shared/models/counter/counter-models';
 import { EventRequest } from '../shared/models/slack/slack-models';
 import { SuppressorService } from '../shared/services/suppressor.service';
 
@@ -46,14 +45,14 @@ async function handleMuzzledMessage(request: EventRequest): Promise<void> {
       request.event.ts,
     );
   } else if (containsTag && (!request.event.subtype || request.event.subtype === 'channel_topic')) {
-    const muzzleId: string | null = await muzzlePersistenceService.getMuzzle(request.event.user, request.team_id);
+    const muzzleId = await muzzlePersistenceService.getMuzzle(request.event.user, request.team_id);
     if (muzzleId) {
       console.log(
         `${userName} attempted to tag someone or change the channel topic. Muzzle increased by ${ABUSE_PENALTY_TIME}!`,
       );
       muzzlePersistenceService.addMuzzleTime(request.event.user, request.team_id, ABUSE_PENALTY_TIME);
       webService.deleteMessage(request.event.channel, request.event.ts, request.event.user);
-      muzzlePersistenceService.trackDeletedMessage(+muzzleId, request.event.text);
+      muzzlePersistenceService.trackDeletedMessage(muzzleId, request.event.text);
       webService
         .sendMessage(
           request.event.channel,
@@ -82,19 +81,23 @@ async function handleBackfire(request: EventRequest): Promise<void> {
       request.team_id,
     );
   } else if (containsTag && (!request.event.subtype || request.event.subtype === 'channel_topic')) {
-    const backfireId = backfireService.getBackfire(request.event.user, request.team_id);
-    console.log(`${userName} attempted to tag someone. Backfire increased by ${ABUSE_PENALTY_TIME}!`);
-    backfireService.addBackfireTime(request.event.user, request.team_id, ABUSE_PENALTY_TIME);
-    webService.deleteMessage(request.event.channel, request.event.ts, request.event.user);
-    backfireService.trackDeletedMessage(+backfireId, request.event.text);
-    webService
-      .sendMessage(
-        request.event.channel,
-        `:rotating_light: <@${request.event.user}> attempted to @ while muzzled! Muzzle increased by ${getTimeString(
-          ABUSE_PENALTY_TIME,
-        )} :rotating_light:`,
-      )
-      .catch(e => console.error(e));
+    const backfireId = await backfireService.getBackfire(request.event.user, request.team_id);
+    if (backfireId) {
+      console.log(`${userName} attempted to tag someone. Backfire increased by ${ABUSE_PENALTY_TIME}!`);
+      backfireService.addBackfireTime(request.event.user, request.team_id, ABUSE_PENALTY_TIME);
+      webService.deleteMessage(request.event.channel, request.event.ts, request.event.user);
+      backfireService.trackDeletedMessage(backfireId, request.event.text);
+      webService
+        .sendMessage(
+          request.event.channel,
+          `:rotating_light: <@${request.event.user}> attempted to @ while muzzled! Muzzle increased by ${getTimeString(
+            ABUSE_PENALTY_TIME,
+          )} :rotating_light:`,
+        )
+        .catch(e => console.error(e));
+    } else {
+      console.log(`Unable to find backfireId for ${request.event.user}`);
+    }
   }
 }
 
@@ -127,25 +130,22 @@ async function handleCounterMuzzle(request: EventRequest): Promise<void> {
 async function handleBotMessage(request: EventRequest, botUserToMuzzle: string): Promise<void> {
   console.log(`A user is muzzled and tried to send a bot message! Suppressing...`);
   webService.deleteMessage(request.event.channel, request.event.ts, request.event.user);
-  const muzzleId: string | null = await muzzlePersistenceService.getMuzzle(botUserToMuzzle, request.team_id);
+  const muzzleId = await muzzlePersistenceService.getMuzzle(botUserToMuzzle, request.team_id);
   if (muzzleId) {
-    muzzlePersistenceService.trackDeletedMessage(+muzzleId, 'A bot message');
+    muzzlePersistenceService.trackDeletedMessage(muzzleId, 'A bot message');
     return;
   }
 
-  const backfireId: string | null = await backfirePersistenceService.getBackfireByUserId(
-    botUserToMuzzle,
-    request.team_id,
-  );
+  const backfireId = await backfirePersistenceService.getBackfireByUserId(botUserToMuzzle, request.team_id);
   if (backfireId) {
-    backfirePersistenceService.trackDeletedMessage(+backfireId, 'A bot user message');
+    backfirePersistenceService.trackDeletedMessage(backfireId, 'A bot user message');
     return;
   }
 
-  const counter: CounterMuzzle | undefined = await counterPersistenceService.getCounterMuzzle(botUserToMuzzle);
-  console.log(counter);
-  if (counter && counter.counterId) {
-    counterPersistenceService.incrementMessageSuppressions(+counter.counterId);
+  const counter = await counterPersistenceService.getCounterMuzzle(botUserToMuzzle);
+
+  if (counter?.counterId) {
+    counterPersistenceService.incrementMessageSuppressions(counter.counterId);
     return;
   }
 }
@@ -189,7 +189,7 @@ eventController.post('/muzzle/handle', async (req: Request, res: Response) => {
     console.time('respond-to-event');
     res.status(200).send();
     const request: EventRequest = req.body;
-    // console.log(request);
+    console.log(request);
     const isNewUserAdded = request.event.type === 'team_join';
     const isNewChannelCreated = request.event.type === 'channel_created';
     const isReaction = request.event.type === 'reaction_added' || request.event.type === 'reaction_removed';
